refactor: 全ての不要な型再エクスポートを削除し、直接インポートに統一
- SetlistFormコンポーネントから不要な後方互換性のための型再エクスポートを削除
- setlist-themes/types.tsファイルを削除し、全て@/types/componentsから直接インポート
- 全ての型インポートを統一された場所から直接取得するように修正
- SongItemInputでSetlistFormItemをSetlistItemとして型エイリアス使用
- CLAUDE.mdのプロジェクト構造からもtypes.tsファイルを削除

完全に削除された型再エクスポート:
- src/components/forms/SetlistForm.tsx の再エクスポート
- src/components/setlist-themes/types.ts ファイル全体

影響ファイル:
- src/app/setlists/[id]/edit/page.tsx
- src/app/setlists/[id]/page.tsx
- src/app/setlists/new/page.tsx
- src/components/forms/SongItemInput.tsx
- src/components/forms/SetlistFormFields.tsx
- src/components/forms/SetlistForm.tsx
- src/components/setlist/ImageGenerator.tsx
- src/components/setlist/SetlistPreview.tsx
- src/components/setlist-themes/BaseTheme.tsx
- src/components/setlist-themes/BlackTheme.tsx
- src/components/setlist-themes/WhiteTheme.tsx
- src/components/setlist-themes/SetlistRenderer.tsx
- src/components/setlist-themes/index.ts
- CLAUDE.md
