+++ src/app/setlists/[id]/edit/page.tsx
@@ -5,7 +5,8 @@ import { Container, Alert, CircularProgress, Typography } from '@mui/material';
 import { useQuery, useMutation } from '@apollo/client';
 import { useParams, useRouter } from 'next/navigation';
 import { GET_SETLIST, UPDATE_SETLIST } from '@/lib/server/graphql/apollo-operations';
-import SetlistForm, { SetlistFormValues } from '@/components/forms/SetlistForm';
+import SetlistForm from '@/components/forms/SetlistForm';
+import { SetlistFormValues } from '@/types/components';
 import { ProtectedRoute } from '@/components/auth/ProtectedRoute';
 
 export default function EditSetlistPage() {
