+++ src/components/setlist-themes/BlackTheme.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BaseTheme } from './BaseTheme';
 import { blackThemeColors } from './themeColors';
-import { SetlistThemeProps } from './types';
+import { SetlistThemeProps } from '@/types/components';
 
 export const BlackTheme: React.FC<SetlistThemeProps> = ({ data, className }) => {
   return <BaseTheme data={data} className={className} colors={blackThemeColors} />;
