+++ src/components/setlist-themes/WhiteTheme.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BaseTheme } from './BaseTheme';
 import { whiteThemeColors } from './themeColors';
-import { SetlistThemeProps } from './types';
+import { SetlistThemeProps } from '@/types/components';
 
 export const WhiteTheme: React.FC<SetlistThemeProps> = ({ data, className }) => {
   return <BaseTheme data={data} className={className} colors={whiteThemeColors} />;
