+++ src/components/setlist-themes/index.ts
@@ -1,4 +1,4 @@
 export { BlackTheme } from './BlackTheme';
 export { WhiteTheme } from './WhiteTheme';
 export { SetlistRenderer } from './SetlistRenderer';
-export type { SetlistData, SetlistItem, SetlistThemeProps } from './types';
+export type { SetlistData, SetlistThemeProps } from '@/types/components';
