--- src/components/setlist-themes/types.ts
@@ -1,3 +1,0 @@
-// 後方互換性のため再エクスポート
-export type { SetlistData, SetlistThemeProps } from '@/types/components';
-export type { SetlistItem } from '@/types/entities';
