+++ src/components/setlist/SetlistPreview.tsx
@@ -3,7 +3,7 @@
 import React, { useState, useEffect } from 'react';
 import { Box, Paper, CircularProgress, Typography, Button } from '@mui/material';
 import QRCode from 'qrcode';
-import { SetlistData } from '../setlist-themes/types';
+import { SetlistData } from '@/types/components';
 import { Theme } from '@/types/common';
 import { SetlistRenderer } from '../setlist-themes/SetlistRenderer';
